refactor : 테스트 코드 수정
seller와 customer 패키지가 user 패키지로 통합 되면서,
모든 Sign 관련 테스트 코드 수정
